Refactor ccompile.h to not include system headers
