+++ include/os/freebsd/linux/types.h
@@ -31,9 +31,6 @@
 #ifndef	_LINUX_TYPES_H_
 #define	_LINUX_TYPES_H_
 
-#include <sys/cdefs.h>
-#include <sys/types.h>
-#include <sys/systm.h>
 #include <linux/compiler.h>
 
 
@@ -53,7 +50,7 @@ typedef uint64_t __le64;
 typedef uint64_t __be64;
 
 typedef unsigned gfp_t;
-typedef uint64_t loff_t;
+typedef off_t loff_t;
 typedef vm_paddr_t resource_size_t;
 typedef uint16_t __bitwise__ __sum16;
 typedef unsigned long pgoff_t;
