+++ module/nvpair/nvpair.c
@@ -30,6 +30,7 @@
 #include <sys/nvpair.h>
 #include <sys/nvpair_impl.h>
 #include <sys/types.h>
+#include <sys/param.h>
 #include <sys/strings.h>
 #include <rpc/xdr.h>
 #include <sys/mod.h>
