+++ module/os/freebsd/spl/spl_kmem.c
@@ -27,6 +27,7 @@
 #include <sys/cdefs.h>
 __FBSDID("$FreeBSD$");
 
+#include <sys/types.h>
 #include <sys/param.h>
 #include <sys/byteorder.h>
 #include <sys/kernel.h>
