+++ module/os/freebsd/spl/spl_string.c
@@ -25,6 +25,7 @@
  * Use is subject to license terms.
  */
 
+#include <sys/types.h>
 #include <sys/param.h>
 #include <sys/string.h>
 #include <sys/kmem.h>
