+++ module/os/freebsd/spl/spl_uio.c
@@ -40,7 +40,7 @@
  * $FreeBSD$
  */
 
-#include <sys/types.h>
+#include <sys/param.h>
 #include <sys/uio.h>
 #include <sys/vnode.h>
